Convert urls relative to current css files to full
If source CSS contain a relative URL not starting with slash (i.e.
relative to current file), those URL will be converted to full relative
path using MEDIA_URL, which should be the same as if COMPRESS is set to
False. When compress enabled, images will be accessible on the same URL.