Respect PIPELINE_AUTO when choosing to compile packages in templatetags.
The documentation states that if PIPELINE_AUTO is off, we won't process
stylesheets and scripts. However, this isn't the case in practice. The
files are still run through a compiler, resulting in new files being
output rather than the originals.

This becomes a problem when trying to use, say, .less files with the
in-browser lessCSS runtime.

The templatetags now respect PIPELINE_AUTO and won't pass anything
through the compilers. This may break some uses, but it adheres to the
documentation and prevents unexpected files when pipelining is turned
off.