Allow compilers to define CSS rel values for uncompiled stylesheets.
When using an uncompiled stylesheet in a different language (such as
less) with an in-browser runtime, it may be important to have a specific
"rel" value. lessCSS's runtime looks for "stylesheet/less", and without
it, the files won't process.

Compilers can now provide an "uncompiled_css_rel" value that will be
used if the compiler matches the filetype and the file isn't being
compiled (due to PIPELINE* being turned off).