Only re-compile LESS files when changed
A new BaseFileTree class is introduced for the compiler to be able to
know the list of files that will be included by the CSS/LESS/Stylus
file, and their modification dates. This it can decide if the output
file needs to be regenerated.

This avoids many costly compilations when working on something else than
the CSS and reloading the page.

Currently only implemented for LESS.