add clarification of how the cached storage works in the docs
I ran into a *very* confusing configuration issue because of this lack
of clarification in these docs and the django docs. Basically I thought
STATIC_ROOT was *just* where django collected the static files to and
then the app server never looked at that directory again, but that is
not the case. Instead, django-pipeline will check the STATIC_ROOT on
requests (with caching). STATIC_ROOT must be configured correctly.

We have nginx serving static files on a separate machine and gunicorn
serving our django app. I compiled the static files on a test machine
and then just rsync'd the files to the nginx server and not
the django server because I thought once all the files were compiled
pipeline would know how to find all the cache-busting filenames. It
turns out it can't and that's because on requests it will look in the
STATIC_ROOT. The django app must be able to find the compiled files in
STATIC_ROOT, which is important if you have separate machines for
deploying, static serving, and/or app serving.