Performance boost in BaseFinderStorage.find_storage
Avoid listing all files for every finder for finding a file. This is
actually a bottleneck when the directory hierarchy is nested or the
number of files served in DEBUG mode is big. Since this flow is done in
every request, it slows down page loading.

In production environments where templates are usually cached performance
doesn't get hit by this.