Exclude node_modules in FileSystemFinder.
Some bower components (such as TinyMCE) require the user to use node
and npm to install build dependencies and build the static
components. This results in a node_modules, created by npm, directory
inside the bower component directory. This is a very large directory
not part of the final built static files. So ignore it.