Simplify execute_command in compiler module
Rewrite thoroughly.

execute command previously caught stdout in memory, which could be expensive.
Now catch in a file.

It also would return stdout to the caller, which would return it to its caller,
which would discard it. Don't return it, now.

It previously interpreted any stderr output as command failure, which is almost
certainly wrong. Warnings go to stderr in most things. Now, only failure to
execute or nonzero return is and error, and Makefiles have a good long history
of enforcing that and only that as a mechanism for detecting errors.

Add docstring.

In verbose output, use repr encoding instead of hand-quoting a dumped string.