Switch to using os.path for outdated file checks in compilers.
The outdated file checks in the compilers were using django storages to check
file existence and compare mtimes, but with my recent change (which restored
the older behavior of comparing the infile and outfile), these were operating
on absolute, local files, which could result in a SuspiciousFileOperation.

The compilers always operate on local files, so we can switch to using the
os.path methods for these instead. This is both more correct and more
efficient.