Fix a regression in determining when to collect static files.
A commit made in 1.6.10 (ae6b3f8eaedffaf09860386c69763ead5a43c94d)
changed a storage lookup to use the prefixed_path variable instead of
the path variable when looking up modification times in order to
determine if an existing collected file should be deleted and
re-collected. By using the wrong path variable, the storage lookups
always failed, resulting in every single file being re-collected every
single time.

This reduces performance considerably, particularly with large codebases.
Since every file ends up copied, they appear as new and are recompiled.
This can result in very long page reloads on a development server,
especially if working on an older machine, in a VM, on a shared
filesystem, or on the Linux subsystem for Windows.

This is a simple change that fixes the lookup to use the correct
variable. Unit tests were added to ensure this continues to work as
expected and does not regress.