Fix string encoding in compiler and compressor errors.
The errors raised when a compiler or compressor failed were including
the message as bytes instead of str. This meant that when printing the
errors, or seeing them in the django-pipeline-error element that's
rendered to the page, the error would be shown as the repr of the
bytestring (which did not properly show newlines, and included extra b''
junk). This change fixes things to decode to str when appropriate.