Ensure correct compiler error styling and strip ANSI escape sequences.
The compiler error output that's injected into the page hard-codes a
background color of white, but didn't hard-code a corresponding text
color. If the page had set a default text color that was too close to
white, it would be hard or impossible to read without selecting the text
on the page.

This change addresses this by hard-coding a text color of black,
ensuring it can be read.

It also goes a step further and improves the display when dealing with
ANSI escape sequences in the error output, making it difficult to read
through or copy/paste meaningful results. We now apply a common regex
for stripping away any typical ANSI escape sequences we should expect in
such output, leaving behind only the plain text content.